no panic for col2im/im2col

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/memory/MemoryManager.java
@@ -78,7 +78,7 @@ public interface MemoryManager {
      *
      * @param windowMillis minimal time milliseconds between calls.
      */
-    void setAutoGcWindow(long windowMillis);
+    void setAutoGcWindow(int windowMillis);
 
     /**
      * This method reutrns
